Move problem report JNI methods to a new module
